week3/pyproject.toml
[project]
name = "week3"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.28.1",
    "massive>=2.0.3",
    "matplotlib>=3.10.8",
    "mcp[cli]>=1.25.0",
    "python-dotenv>=1.2.1",
    "yfinance>=1.0",
]
